refactor: Update inference.py to handle either a single image directory or a list of image directories
This commit modifies the `inferece_dust3r` function in `inference.py` to accept either a single image directory or a list of image directories as the `image_dir_or_list` parameter. If a list is provided, the function will load images from each directory. If a single directory is provided, the function will load images from that directory. This change improves the flexibility of the function and allows for more versatile usage.